requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
pdf2docx
Pillow
aiofiles
pymupdf